+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - `rateLimit` - Added `maxItemsPerPeriod` option along with the options from `throughputLimiter`.
 - `throughputLimiter`
-    - Added `minWindowLength`, `period`, and `expireAfter` options.
+- Added `minWindowLength`, `period`, `expireAfter`, and `getTimeframe` options.
     - The `windowLength` option is now `maxWindowLength`.
     - Separated `throttle` and `append`. These two convenience methods were also created: `throttleAndAppend` and `appendAndThrottle`.